requirements.txt
alembic==1.3.0
aniso8601==8.0.0
attrs==20.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
cryptography==3.2
dominate==2.4.0
environs==8.0.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.14
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
more-itertools==8.5.0
oauthlib==3.1.0
objcrypt==0.6
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pymsteams==0.1.14
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.12.0
SQLAlchemy==1.3.8
toml==0.10.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==0.16.0
zipp==3.1.0
